REF: numeric index test_indexing
